Add files via upload
Q-Learning algorithm applied to plan robot path to reach target position and avoid obstacle. The algorithm generate a Q table that provides quality of reach action at particular state space. The greedy policy chose the highest quality value to select an action. If target and obstacle position remained static, the Q-table can be reused no matter of where is the starting position of the robot end-effector. Once they changed, the Q-learning has to be re-run again to generate a new Q-table. In reality, the obstacles and targets are generally dynamic.