Add files via upload
Deep Q-Learning Network (DQN) is applied to see how deep learning and Q-learning RL can be combined to learn the dynamics of random positions of 2D grid world - relationships among agent - robot, obstacle, and target. The DQN used neural network as the Q function to learn and generate the actions, instead of a Q-table. Instead of state space of only starting position in Q-Learning, DQN defined 3 tensors of starting, obstacle, and target states. Once learned, the DQN is able to plan robot paths with random positions of start, target and obstacle, which is significantly different from the Q-Learning.